+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ### 新特性
 * 【core】       增加MapUtil.removeAny（issue#612@Github）
+* 【core】       Convert.toList支持[1,2]字符串（issue#I149XN@Gitee）
+* 【core】       修正DateUtil.thisWeekOfMonth注释错误（issue#614@Github）
+* 【core】       DateUtil增加toLocalDate等方法，DateTime更好的支持时区
 
 ### Bug修复
 * 【db】         修复MetaUtil.getTableMeta()方法未释放ResultSet的bug（issue#I148GH@Gitee）